Last changes before the deadline

+++ app/src/main/java/com/example/chess/Bishop.java
@@ -1,7 +1,7 @@
 package com.example.chess;
 
 public class Bishop extends Piece {
-    private int value = 350;
+    private double value = 350;
 
     public Bishop(boolean colour) {
         super(colour);
@@ -12,7 +12,7 @@ public class Bishop extends Piece {
     }
 
     @Override
-    public int getValue()
+    public double getValue()
     {
         return value;
     }
